services/development/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pdfplumber==0.11.0
google-generativeai==0.3.2
python-dateutil==2.8.2
email-validator==2.1.0
aiofiles==23.2.1
pillow==10.2.0
